+++ gravitee-node-cluster/gravitee-node-cluster-plugin-handler/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,5 @@
 io.gravitee.plugin.core.api.PluginHandler=\
-io.gravitee.node.cluster.plugin.ClusterPluginHandler
+io.gravitee.node.cluster.plugin.ClusterPluginHandler,\
+io.gravitee.node.cluster.plugin.HazelcastProviderPluginHandler
 io.gravitee.node.api.healthcheck.Probe=\
     io.gravitee.node.cluster.healthcheck.ClusterProbe
