+++ gravitee-node-cluster/gravitee-node-cluster-plugin-hazelcast-provider/src/main/resources/plugin.properties
@@ -1,0 +1,6 @@
+id=hazelcast-provider
+name=Hazelcast Provider
+version=${project.version}
+description=${project.description}
+class=io.gravitee.node.plugin.cluster.hazelcast.provider.HazelcastProvider
+type=hazelcast-provider
