+++ gravitee-node-cluster/pom.xml
@@ -34,6 +34,7 @@
     <modules>
         <module>gravitee-node-cluster-plugin-handler</module>
         <module>gravitee-node-cluster-plugin-hazelcast</module>
+        <module>gravitee-node-cluster-plugin-hazelcast-provider</module>
         <module>gravitee-node-cluster-plugin-standalone</module>
     </modules>
 </project>
